Improve migrations

+++ db/migrate/20140530115433_drop_badges_table.rb
@@ -1,6 +1,6 @@
 class DropBadgesTable < ActiveRecord::Migration
   def up
-    drop_table :badges
+    drop_table :badges if ActiveRecord::Base.connection.table_exists? 'badges'
   end
 
   def down
